+++ src/main/java/app/controller/Controller.java
@@ -1,20 +1,21 @@
 package app.controller;
 
+import java.util.Objects;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.stage.Stage;
-import java.util.Objects;
 
 abstract class Controller {
 
   protected Stage stage;
-  private final static String ERROR = "Failed to load .fxml file";
+  private static final String ERROR = "Failed to load .fxml file";
   private static final String ALERT_HEADER = "Error";
 
   /**
-   * Constructor for controllers, takes current @param stage and @param fxmlpath and switches to next view upon creation.
+   * Constructor for controllers,
+   * takes current @param stage and @param fxmlpath and switches to next view upon creation.
    */
   Controller(Stage stage, String fxmlpath) {
     try {