+++ src/main/java/protocol/Response.java
@@ -13,7 +13,7 @@ public class Response {
   public static class Types {
     public static final String ERROR = "error";
     public static final String LOBBY_UPDATE = "lobby update";
-    public static final String GAME_STARTING = "game starting";
+    public static final String GAME_STARTING = "race starting";
     public static final String REGISTERED = "registered";
     public static final String OPEN_LOBBIES = "open lobbies";
     public static final String RACE_UPDATE = "race update";